# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tender Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tenders
    description: Tender management operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tenders:
    get:
      operationId: listTenders
      tags: [tenders]
      summary: List all tenders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: entity
          in: query
          schema:
            type: string
        - name: urgent
          in: query
          schema:
            type: boolean
        - name: won
          in: query
          schema:
            type: boolean
        - name: engineer
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tenders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tender"
    post:
      operationId: createTender
      tags: [tenders]
      summary: Create a new tender
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenders/stats:
    get:
      operationId: getTenderStats
      tags: [tenders]
      summary: Get tender statistics
      responses:
        "200":
          description: Tender statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderStats"
  /tenders/{id}:
    get:
      operationId: getTender
      tags: [tenders]
      summary: Get a tender by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTender
      tags: [tenders]
      summary: Update a tender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTender
      tags: [tenders]
      summary: Delete a tender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Tender:
      type: object
      required:
        - id
        - tenderNumber
        - projectName
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        tenderNumber:
          type: string
        governmentEntity:
          type: ["string", "null"]
        projectName:
          type: string
        tenderType:
          type: ["string", "null"]
        announcementDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        preliminaryMeetingHeld:
          type: boolean
        preliminaryMeetingDate:
          type: ["string", "null"]
        bondValue:
          type: ["number", "null"]
        docsValue:
          type: ["number", "null"]
        responsibleEngineer:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - new
            - studying
            - requesting_quotes
            - preparing_technical
            - preparing_financial
            - management_review
            - ready_to_submit
            - submitted
            - under_evaluation
            - won
            - lost
            - cancelled
        offerValue:
          type: ["number", "null"]
        profitPercentage:
          type: ["number", "null"]
        isSubmitted:
          type: boolean
        winner:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        fileConditions:
          type: ["string", "null"]
        filePricing:
          type: ["string", "null"]
        fileSuppliers:
          type: ["string", "null"]
        fileOpening:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TenderInput:
      type: object
      required:
        - tenderNumber
        - projectName
        - status
      properties:
        tenderNumber:
          type: string
          minLength: 1
        governmentEntity:
          type: string
        projectName:
          type: string
          minLength: 1
        tenderType:
          type: string
        announcementDate:
          type: string
        deadline:
          type: string
        preliminaryMeetingHeld:
          type: boolean
        preliminaryMeetingDate:
          type: string
        bondValue:
          type: number
        docsValue:
          type: number
        responsibleEngineer:
          type: string
        status:
          type: string
          enum:
            - new
            - studying
            - requesting_quotes
            - preparing_technical
            - preparing_financial
            - management_review
            - ready_to_submit
            - submitted
            - under_evaluation
            - won
            - lost
            - cancelled
        offerValue:
          type: number
        profitPercentage:
          type: number
        isSubmitted:
          type: boolean
        winner:
          type: string
        notes:
          type: string
        fileConditions:
          type: string
        filePricing:
          type: string
        fileSuppliers:
          type: string
        fileOpening:
          type: string
    TenderUpdate:
      type: object
      properties:
        tenderNumber:
          type: string
          minLength: 1
        governmentEntity:
          type: string
        projectName:
          type: string
          minLength: 1
        tenderType:
          type: string
        announcementDate:
          type: string
        deadline:
          type: string
        preliminaryMeetingHeld:
          type: boolean
        preliminaryMeetingDate:
          type: string
        bondValue:
          type: number
        docsValue:
          type: number
        responsibleEngineer:
          type: string
        status:
          type: string
          enum:
            - new
            - studying
            - requesting_quotes
            - preparing_technical
            - preparing_financial
            - management_review
            - ready_to_submit
            - submitted
            - under_evaluation
            - won
            - lost
            - cancelled
        offerValue:
          type: number
        profitPercentage:
          type: number
        isSubmitted:
          type: boolean
        winner:
          type: string
        notes:
          type: string
        fileConditions:
          type: string
        filePricing:
          type: string
        fileSuppliers:
          type: string
        fileOpening:
          type: string
    TenderStats:
      type: object
      required:
        - total
        - byStatus
        - urgentCount
        - wonCount
        - totalOfferValue
        - winRate
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        urgentCount:
          type: integer
        wonCount:
          type: integer
        totalOfferValue:
          type: number
        winRate:
          type: number
    StatusCount:
      type: object
      required:
        - status
        - count
      properties:
        status:
          type: string
        count:
          type: integer
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file.
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path. Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string